requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.2
django-environ==0.10.0
idna==3.4
Pillow==9.5.0
sqlparse==0.4.4
stripe==5.5.0
tzdata==2023.3
urllib3==2.0.3

django-allauth~=0.54.0